requirements.txt
streamlit
transformers
sentence-transformers
scikit-learn
faiss-cpu
numpy
pandas
torch
plotly


